First commit of the code exported from Skyman
This is the first Kamino project commit of the VMSS Prototype Pattern
code extracated from the Skyman project.  I did a quick hack and
pull together of the required bits and put it into this one python
file.  This removed the dependency on many of the Skyman "scm" tool
behaviors and libraries but had to cut-n-paste a few into here
to replace needed elements.

I have tested this minimally to work given the right inputs.
More details need to be worked out and the code cut down some
more as it is still structured to work for the Skyman system
(with many bits removed)

There will be some additional work to make it more generic and
document it outside of the Skyman structure.