Lots of cleanups to remove more Skyman platform structures
The big one is that since we give a target node (it is required),
we know which VMSS we will be dealing with during update.

Note that the status shows all VMSS instances and their status.

Changed the subscription to be optional - such that in an environment
where you are logged into the default subscription you don't need to
provide that anymore.

Added more validation of the target-node parameter.

Removed the dependency on rebooter to clean up an annotation.  We
now do that after the snapshot has been taken (or after a failure)

Added an option as to if drian failure should be ignored.  Normally
one would not want to ignore this but some people may not be in a
situation in which drain doe succeed.  This makes sure that drain
failure can be acceptable (--force)

If target node does not exist, fail early on.  Still needs to deal
with it not existing later but an early and fast fail is still better
for poorly entered node name.

Changed the way we delete extensions to be one of deleting all but
the billing extension.  In fact, I think I will do the same on our
clusters now - the reason is that semantically we really want all
extensions deleted.  The only reason the billing extension is kept
is that aks-engine uses that get signal as to new nodes in clusters.
It is a no-op code wise, it is just how many times it is run.

Added more uses of the retry_func to stop retries if we get the
"NotFound" error type - as retrying those does not help.