First work on the Helm chart for Kamino VMSS Prototype
This is the first bit of work for the helm chart and related
changes to the code to make it better to work in a cluster.

However, in working on this I have come to realize that the cluster
name should not be needed - especially since the azure.json file
does not contain the cluster name (but it does contain the resource
group name)

As such, some refactoring to remove the cluster name and some other
changes to the code.

The new smoketest2.sh is just there to show what use would look like.
It is currently very much tied to my cluster and my test acr in skyman.
So I am sure it will not work for you.  It is more as documentation
than actually useful.

This has been tested in my clusters for both status and actual image
creation.

PS - I merged in the Kamino branding changes as there would have been
a lot of problems with the merge later if I did not.