Add --maximum-image-age option
This adds an option called --maximum-image-age and it defines
the number days old an image can get before the auto-selection
process no longer depends on only new OS patches.

This can be used to make sure refreshed images are picked that
may have updated docker caches or other state that is not tracked
by the OS patches.

We use this to allow our images to stay more "fresh" with respect
to the workload running in the cluster.

The default is 0, which means that only new OS patches will trigger
and new image to be created.  The value is in days and anything
less than 1 would effectively meaningles anyway.

Note that if there was an OS patch and the image age is too old,
we still prefer to pick up the patched nodes over the unpatched
nodes, as long as the number of patched nodes satisfies the
minimum candidates value.