Add filtering to prevent masters from being selected
Normally masters are not part of a VMSS but there is a feature
in aks-engine to enable masters on a VMSS.  However, masters can
not just scale in and out so there is (a) no good reason to make
a VMSS Prototype for them and (b) we can't safely do it anyway.

This just makes sure that we filter out any master which means
that even if it is within a VMSS, they are never selected and thus
never get an image built.

Fixes #48