Add kamino.auto.crontab feature
Kamino is designed to be lights-out operation.  As such, it can
be installed as a crontab that runs the auto-update process on
a regular basis.

In most of our clusters, we run this automatic update process on
a daily basis.  Since OS updates usually don't happen daily, this
ends up just checking and noting that there is nothing to do.

When an OS update does happen (or the age limit on the prior image
is reached) then the auto-update process does what it needs to do.

This allows the whole thing to be running in clusters without
intervention.

The schedule could easily be run more often as the cronjob is set
up to not run while a prior instance is running.  I run my test
clusters with this turned to running every few minutes just to
stress test the logic and to quickly notice a condition where a
new image is needed (just to reduce testing time).

This addresses issue #44

Enhanced smoketest3.sh a bit to make it do what I needed.

The new smoketest4.sh is just like smoketest3.sh only it uses the
cronjob (with an "every minute" schedule) to run the cronjob and
then uninstalls after it completes so as to not continually rerun.
It is just a smoketest of the chart...