Add some code to deal with the machine-id
This fixes issue #22 by zapping the /etc/machine-id such that
new nodes, on boot, will define their own unique value rather than
retain the value from the prior instance.

This required a few extra helm chart changes.

Since this is not strictly required, we just "try" to do this but
if it fails, we move on and create the disk image like we did before.