Add kamino.auto.crontab feature (#62)
* Add kamino.auto.crontab feature

Kamino is designed to be lights-out operation.  As such, it can
be installed as a crontab that runs the auto-update process on
a regular basis.

In most of our clusters, we run this automatic update process on
a daily basis.  Since OS updates usually don't happen daily, this
ends up just checking and noting that there is nothing to do.

When an OS update does happen (or the age limit on the prior image
is reached) then the auto-update process does what it needs to do.

This allows the whole thing to be running in clusters without
intervention.

The schedule could easily be run more often as the cronjob is set
up to not run while a prior instance is running.  I run my test
clusters with this turned to running every few minutes just to
stress test the logic and to quickly notice a condition where a
new image is needed (just to reduce testing time).

This addresses issue #44

Enhanced smoketest3.sh a bit to make it do what I needed.

The new smoketest4.sh is just like smoketest3.sh only it uses the
cronjob (with an "every minute" schedule) to run the cronjob and
then uninstalls after it completes so as to not continually rerun.
It is just a smoketest of the chart...

* Added some extra help for the cronjob schedule value

* Add some code to deal with the machine-id

This fixes issue #22 by zapping the /etc/machine-id such that
new nodes, on boot, will define their own unique value rather than
retain the value from the prior instance.

This required a few extra helm chart changes.

Since this is not strictly required, we just "try" to do this but
if it fails, we move on and create the disk image like we did before.

* Forgot to remove an entry that is only for my clusters

* Forgot another skyman cluster specific setting...

* Clean up the logging a bit

Also add a longer timeout to the smoketest, just in case there
is some extended period of image pull

* Update helm/vmss-prototype/README.md

Co-authored-by: Amr Hanafi <amramahdi@gmail.com>

* minor formatting

* Some enhanced wording describing the cronjob benefit

* Make the mount even more restricted

This makes the tweaks mount restricted to the file /etc/machine-id
such that it only has access to that.  A much cleaner mount.

Co-authored-by: Amr Hanafi <amramahdi@gmail.com>