A full fix for issue #22 - machine-id
The prior fix only helped by making a new machine-id due to the
way the vm deallocation hapened in Azure and the OS setup on the
machine.

This required a bit more complex operation to shutdown the OS
right after the machine-id was cleared.  The shutdown of the OS
works nicely and safely in that all buffers/files are closed/flushed
and the disk is "synced" before issuing the power-off.  But it does
not let the services do a re-initialization of the machine-id

This same process is now in production for our Skyman based clusters
and has resulted in a few thousand new VMs scaling in with unique
machine-id values.