Address the delete of the donor instance
The donor instance no longer is deleted when creating the prototype
image.  This was always the plan, to not delete the donor, but there
were two problems with it.

The first, and most important, was issue #26 which is now addressed.
The problem was that the cloud-init data "knew" the node was a different
name in a prior instance and thus registered a DNS name delete of its
"old" name.  This, unfortunately would remove the DNS entry for the
donor node.  The entry would come back when the donor node restarts or
otherwise reruns its DDNS code but since every scaled in node with that
donor node's instance would do that, it would continually cause problems
for the donor.  This has been addressed with this change.

Another issue was that when certain strong VMSS restarts are done,
the cloud-init may rerun and certain versions of aks-engine had set
up cloud-init data to zap the azure.json file.  This could only impact
the donor node and one if the donor node was not yet a node built by
vmss-prototype (so, basically the first one for each pool).

Newer aks-engine versions don't have this problem and since it is a
"one-time" type of problem, we can just ignore it as most do not run
the problem aks-engine versions.

I got some of the documentation updated with respect to the new
behavior.  We will need to do a run through all of it to see
if there are any other changes.

One new feature is that we now track the ancestry of the image.
Each time we create an image, we append to the /var/log/ancestry.log
a line with the timestamp and node name that we are building the
image with.  This thus has the genetic heritage of the node
image.