Add a job TTL such that they don't hang around forever
After a job completes, the pod does not need to stay on the cluster
as a "completed" pod forever.  It just takes up kubernetes state
space when it has long since completed.  Keeping it around for a
while after completing allows for inspection but once that time
has been reached, it should evaporate and no longer take up state
space.  (Jobs don't restart once completed so they are not very
useful and in larger clusters the state space becomes a bottleneck)