Change to "stop" from "deallocate"
Stop works noticably faster in Azure as it does not deallocate
the NIC or IP addresses associated with the VM.  This saves
also saves significant time in restarting the VM after it was
stopped.

We stopped doing this to reduce the node down time to get the
snapshot made and reduce the potential failure paths within
Azure as deallocate and restart involves so many more moving
pieces.