Add workaround for observed strange behavior
We had a test run that observed an extension still being configured
even after it was deleted successfully from the configuration.

This was reported in issue/bug #83

To hopefully address this, we re-list the extensions after trying
to delete them and will re-delete those that we feel we should
delete if they still are there after having deleted them.  We
will do that up to 5 times.

The problem is that at this point the new image has been installed
and set up as the target image so the extensions should not be needed
but we can't undo what we did to install new image.  We can only go
forward.  (There is no single operation that can do it all atomically,
especially since now we see that even a successful delete is not
always successful anyway)