Minor change to work around an azure CLI bug in 2.30.0
The latest Azure CLI has a bug where it break the snapshot
command when run against a VMSS disk instance.  Turns out that
there is a simple workaround that is perfectly fine to use as
it just forces the path that we normally would have gone down
anyway.

See https://github.com/Azure/azure-cli/issues/20313