Address the potential buildup of journal logs (#95)
When a new node is created from a prototype image, the journal
logs are generated in a new, unique hashed directory.  This,
unfortunately can no longer match the prior directory name
and the prior log files are then not cleaned up by the
journalctl vacuum processes.  Thus, this could lead to leaking
of disk space over time and multiple generations of prototype images.